Adding specs for the resource downloader and download handler
